🚚 rename: api 파일 구조 및 파일명 수정 (#31)
