+++ src/api/artistApi.ts
@@ -33,54 +33,6 @@ export const getArtistTopTracks = async (artist_id: any) => {
   }
 }
 
-export const getRecommendations = async (artists, genre, tracks) => {
-  try {
-    const response = await baseInstance(
-      `/recommendations?limit=8&market=KR&seed_artists=${artists}&seed_genres=${genre}&seed_tracks=${tracks}`,
-      {
-        headers: {
-          Authorization: `Bearer ${localStorage.getItem('non-member-token')}`
-        }
-      }
-    )
-    const data = response.data
-    const { tracks: responseTrack } = data
-    return responseTrack
-  } catch (error) {
-    console.error(error)
-  }
-}
-
-export const getTracksAudioFeatures = async (tracksId: string) => {
-  try {
-    const response = await baseInstance(`audio-features?ids=${tracksId}`, {
-      headers: {
-        Authorization: `Bearer ${localStorage.getItem('non-member-token')}`
-      }
-    })
-    const data = response.data
-    const { audio_features } = data
-    return audio_features
-  } catch (error) {
-    console.error(error)
-  }
-}
-
-export const getPlaylistTop50 = async (playlist_id: string) => {
-  try {
-    const response = await baseInstance(`playlists/${playlist_id}`, {
-      headers: {
-        Authorization: `Bearer ${localStorage.getItem('non-member-token')}`
-      }
-    })
-    const data = response.data
-    const trackList = data.tracks.items
-    return { trackList }
-  } catch (error) {
-    console.error(error)
-  }
-}
-
 export const getArtistAlbums = async (artist_id: string) => {
   try {
     const response = await baseInstance(
@@ -96,6 +48,7 @@ export const getArtistAlbums = async (artist_id: string) => {
     return { albumList }
   } catch (error) {
     console.error(error)
+    return error
   }
 }
 
@@ -114,19 +67,6 @@ export const getRelatedArtists = async (artist_id: string) => {
     return { artists }
   } catch (error) {
     console.error(error)
-  }
-}
-
-export const getAlbum = async (album_id: string) => {
-  try {
-    const response = await baseInstance(`albums/${album_id}/?market=KR`, {
-      headers: {
-        Authorization: `Bearer ${localStorage.getItem('non-member-token')}`
-      }
-    })
-    const data = response.data
-    return data
-  } catch (error) {
-    console.error(error)
+    return error
   }
 }
