+++ src/api/recommendApi.ts
@@ -1,0 +1,19 @@
+import { baseInstance } from './axios'
+
+export const getRecommendations = async (artists, genre, tracks) => {
+  try {
+    const response = await baseInstance(
+      `/recommendations?limit=8&market=KR&seed_artists=${artists}&seed_genres=${genre}&seed_tracks=${tracks}`,
+      {
+        headers: {
+          Authorization: `Bearer ${localStorage.getItem('non-member-token')}`
+        }
+      }
+    )
+    const data = response.data
+    const { tracks: responseTrack } = data
+    return responseTrack
+  } catch (error) {
+    console.error(error)
+  }
+}
