+++ src/assets/icons/SpotifyHover.tsx
@@ -1,5 +1,3 @@
-import React from "react";
-
 export const SpotifyHover = () => {
   return (
     <svg
@@ -7,12 +5,34 @@ export const SpotifyHover = () => {
       height="38"
       viewBox="0 0 265 38"
       fill="none"
-      xmlns="http://www.w3.org/2000/svg"
-    >
-      <rect x="3" y="3" width="259" height="27" fill="white" />
-      <rect x="3" y="3" width="3" height="3" fill="white" />
-      <rect y="6" width="3" height="20" fill="white" />
-      <rect x="3" y="26" width="3" height="3" fill="white" />
+      xmlns="http://www.w3.org/2000/svg">
+      <rect
+        x="3"
+        y="3"
+        width="259"
+        height="27"
+        fill="white"
+      />
+      <rect
+        x="3"
+        y="3"
+        width="3"
+        height="3"
+        fill="white"
+      />
+      <rect
+        y="6"
+        width="3"
+        height="20"
+        fill="white"
+      />
+      <rect
+        x="3"
+        y="26"
+        width="3"
+        height="3"
+        fill="white"
+      />
       <rect
         x="262"
         y="29"
@@ -37,15 +57,44 @@ export const SpotifyHover = () => {
         transform="rotate(-180 262 6)"
         fill="white"
       />
-      <rect x="128" y="29" width="9" height="6" fill="white" />
-      <rect x="131" y="29" width="3" height="9" fill="white" />
-      <rect x="6" y="29" width="40" height="3" fill="white" />
-      <rect x="41" y="29" width="218" height="3" fill="white" />
-      <rect x="6" width="253" height="3" fill="white" />
+      <rect
+        x="128"
+        y="29"
+        width="9"
+        height="6"
+        fill="white"
+      />
+      <rect
+        x="131"
+        y="29"
+        width="3"
+        height="9"
+        fill="white"
+      />
+      <rect
+        x="6"
+        y="29"
+        width="40"
+        height="3"
+        fill="white"
+      />
+      <rect
+        x="41"
+        y="29"
+        width="218"
+        height="3"
+        fill="white"
+      />
+      <rect
+        x="6"
+        width="253"
+        height="3"
+        fill="white"
+      />
       <path
         d="M17 16.25V17H11V16.25H10.25V15.5H9.5V14H10.25V13.25H11V12.5H17V13.25H17.75V14H18.5V15.5H17.75V16.25H17ZM11.75 13.25V14H11V15.5H11.75V16.25H16.25V15.5H17V14H16.25V13.25H11.75ZM11 21.5V18.5H8.75V17.75H19.25V18.5H17V21.5H15.5V18.5H12.5V21.5H11ZM22.25 17H21.5V14.75H29V13.25H21.5V12.5H30.5V15.5H23V17H30.5V17.75H29V20H31.25V20.75H20.75V20H23V17.75H22.25V17ZM27.5 20V17.75H24.5V20H27.5ZM41 17.75H40.25V17H39.5V16.25H38.75V14.75H39.5V14H40.25V13.25H38.75V12.5H41V11.75H42.5V12.5H44.75V13.25H43.25V14H44V14.75H44.75V16.25H44V17H43.25V17.75H42.5V20H47V11.75H48.5V21.5H47V20.75H38.75V20H41V17.75ZM42.5 14H41V14.75H40.25V16.25H41V17H42.5V16.25H43.25V14.75H42.5V14ZM53 16.25H50.75V15.5H52.25V14.75H50.75V12.5H52.25V11.75H55.25V12.5H56.75V14.75H55.25V15.5H59V11.75H60.5V17.75H59V17H56.75V16.25H54.5V17.75H53V16.25ZM54.5 13.25V12.5H53V13.25H52.25V14H53V14.75H54.5V14H55.25V13.25H54.5ZM51.5 20.75V17.75H53V20.75H60.5V21.5H52.25V20.75H51.5ZM72.5 11.75V21.5H71V11.75H72.5ZM67.25 20V20.75H64.25V20H63.5V19.25H62.75V14H63.5V13.25H64.25V12.5H67.25V13.25H68V14H68.75V19.25H68V20H67.25ZM65 13.25V14H64.25V19.25H65V20H66.5V19.25H67.25V14H66.5V13.25H65ZM83.75 11.75V16.25H86V17H83.75V21.5H82.25V11.75H83.75ZM79.25 20V20.75H75.5V20H74.75V15.5H78.5V13.25H74.75V12.5H80V16.25H76.25V20H78.5V19.25H80V20H79.25ZM92 17H87.5V16.25H86.75V11.75H92.75V13.25H95V11.75H96.5V17H95V15.5H92.75V16.25H92V17ZM91.25 16.25V12.5H88.25V16.25H91.25ZM95 14.75V14H92.75V14.75H95ZM89 17.75V20.75H96.5V21.5H88.25V20.75H87.5V17.75H89ZM110.75 16.25V17H109.25V16.25H108.5V15.5H107V16.25H106.25V17H104.75V16.25H105.5V15.5H106.25V14.75H107V12.5H104.75V11.75H110.75V12.5H108.5V14.75H109.25V15.5H110V16.25H110.75ZM114.5 11.75V17H113V14.75H110.75V14H113V11.75H114.5ZM107 17.75V20.75H114.5V21.5H106.25V20.75H105.5V17.75H107ZM122 19.25V20H122.75V20.75H121.25V20H120.5V19.25H119V20H118.25V20.75H116.75V20H117.5V19.25H118.25V18.5H119V13.25H116.75V12.5H119V11.75H120.5V12.5H122.75V13.25H120.5V18.5H121.25V19.25H122ZM125 11.75V21.5H123.5V17H122V16.25H123.5V11.75H125ZM127.25 11.75V21.5H125.75V11.75H127.25ZM134.75 16.25H136.25V15.5H137V12.5H129.5V11.75H138.5V15.5H137.75V16.25H139.25V17H128.75V16.25H133.25V14H134.75V16.25ZM129.5 18.5V17.75H138.5V20.75H137.75V21.5H137V18.5H129.5ZM141.5 14.75V12.5H143V11.75H149V12.5H150.5V14.75H149V15.5H143V14.75H141.5ZM143.75 12.5V13.25H143V14H143.75V14.75H148.25V14H149V13.25H148.25V12.5H143.75ZM151.25 16.25V17H140.75V16.25H151.25ZM141.5 20.75V19.25H149V18.5H141.5V17.75H150.5V20H143V20.75H150.5V21.5H142.25V20.75H141.5ZM159.5 14.75V11.75H168.5V12.5H161V14.75H168.5V15.5H160.25V14.75H159.5ZM169.25 16.25V17H158.75V16.25H169.25ZM159.5 20.75V19.25H167V18.5H159.5V17.75H168.5V20H161V20.75H168.5V21.5H160.25V20.75H159.5ZM171.5 14.75V12.5H173V11.75H179V12.5H180.5V14.75H179V15.5H173V14.75H171.5ZM173.75 12.5V13.25H173V14H173.75V14.75H178.25V14H179V13.25H178.25V12.5H173.75ZM181.25 16.25V17H170.75V16.25H181.25ZM171.5 20.75V19.25H179V18.5H171.5V17.75H180.5V20H173V20.75H180.5V21.5H172.25V20.75H171.5ZM195.5 17V16.25H194.75V15.5H193.25V16.25H192.5V17H189.5V16.25H191.75V15.5H192.5V14.75H193.25V12.5H194.75V14.75H195.5V15.5H196.25V16.25H198.5V17H195.5ZM199.25 17.75V18.5H194.75V21.5H193.25V18.5H188.75V17.75H199.25ZM211.25 16.25V17H208.25V16.25H207.5V15.5H206.75V13.25H207.5V12.5H208.25V11.75H211.25V12.5H212V13.25H212.75V15.5H212V16.25H211.25ZM216.5 11.75V17H215V11.75H216.5ZM209 12.5V13.25H208.25V15.5H209V16.25H210.5V15.5H211.25V13.25H210.5V12.5H209ZM209 20.75V21.5H207.5V20.75H208.25V20H209V17.75H210.5V20H211.25V20.75H212V21.5H210.5V20.75H209ZM214.25 20.75V21.5H212.75V20.75H213.5V20H214.25V17.75H215.75V20H216.5V20.75H217.25V21.5H215.75V20.75H214.25ZM224 20H223.25V20.75H220.25V20H219.5V19.25H218.75V14H219.5V13.25H220.25V12.5H223.25V13.25H224V14H224.75V16.25H227V11.75H228.5V21.5H227V17H224.75V19.25H224V20ZM221 19.25V20H222.5V19.25H223.25V14H222.5V13.25H221V14H220.25V19.25H221ZM232.25 16.25H231.5V14H232.25V13.25H233V12.5H239V13.25H239.75V14H240.5V16.25H239.75V17H239V20H241.25V20.75H230.75V20H233V17H232.25V16.25ZM238.25 14V13.25H233.75V14H233V16.25H233.75V17H238.25V16.25H239V14H238.25ZM237.5 20V17.75H234.5V20H237.5ZM249.5 18.5V17H250.25V16.25H251.75V15.5H248.75V14.75H251.75V14H248.75V14.75H246.5V15.5H248.75V16.25H246.5V19.25H245.75V20H244.25V19.25H243.5V17.75H244.25V17H245.75V13.25H248.75V12.5H252.5V18.5H251.75V19.25H250.25V18.5H249.5Z"
         fill="black"
       />
     </svg>
-  );
-};
+  )
+}
